add eslint Fixes #5
