+++ .prettierrc.json
@@ -1,0 +1,8 @@
+{
+    "arrowParens": "avoid",
+    "jsxBracketSameLine": true,
+    "semi": false,
+    "singleQuote": true,
+    "tabWidth": 4,
+    "trailingComma": "all"
+}
