+++ index.html
@@ -2,38 +2,38 @@
 <html lang="en">
 
 <head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-  <link rel="apple-touch-icon" sizes="180x180" href="/assets/favicon/apple-touch-icon.png">
-  <link rel="icon" type="image/png" sizes="32x32" href="/assets/favicon/favicon-32x32.png">
-  <link rel="icon" type="image/png" sizes="16x16" href="/assets/favicon/favicon-16x16.png">
-  <link rel="manifest" href="/assets/favicon/site.webmanifest">
-  <link rel="mask-icon" href="/assets/favicon/safari-pinned-tab.svg" color="#5bbad5">
-  <meta name="theme-color" content="#333333">
+    <link rel="apple-touch-icon" sizes="180x180" href="/assets/favicon/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/assets/favicon/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/assets/favicon/favicon-16x16.png">
+    <link rel="manifest" href="/assets/favicon/site.webmanifest">
+    <link rel="mask-icon" href="/assets/favicon/safari-pinned-tab.svg" color="#5bbad5">
+    <meta name="theme-color" content="#333333">
 
-  <!-- link website preview -->
-  <meta property="og:title" content="Sensasi Delight: Adam &quot;Zain&quot; Akbar" />
-  <meta property="og:description" content="Full time learner, mostly web developer." />
-  <meta property="og:image" content="/assets/images/web-preview.jpg" />
-  <meta property="og:url" content="https://sensasi-delight.web.app" />
+    <!-- link website preview -->
+    <meta property="og:title" content="Sensasi Delight: Adam &quot;Zain&quot; Akbar" />
+    <meta property="og:description" content="Full time learner, mostly web developer." />
+    <meta property="og:image" content="/assets/images/web-preview.jpg" />
+    <meta property="og:url" content="https://sensasi-delight.web.app" />
 
-  <!-- link website preview -->
-  <meta name="twitter:card" content="summary_large_image" />
-  <meta name="twitter:site" content="@sensasi_delight" />
-  <meta name="twitter:creator" content="@sensasi_delight" />
-  <meta name="twitter:title" content="Sensasi Delight: Adam &quot;Zain&quot; Akbar" />
-  <meta name="twitter:description" content="Full time learner, mostly web developer." />
-  <meta name="twitter:image" content="/assets/images/web-preview.jpg" />
+    <!-- link website preview -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="@sensasi_delight" />
+    <meta name="twitter:creator" content="@sensasi_delight" />
+    <meta name="twitter:title" content="Sensasi Delight: Adam &quot;Zain&quot; Akbar" />
+    <meta name="twitter:description" content="Full time learner, mostly web developer." />
+    <meta name="twitter:image" content="/assets/images/web-preview.jpg" />
 
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
 
-  <title>Sensasi Delight: Adam "Zain" Akbar</title>
+    <title>Sensasi Delight: Adam "Zain" Akbar</title>
 </head>
 
 <body>
-  <div id="root"></div>
-  <script type="module" src="/src/main.jsx"></script>
+    <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
 </body>
 
 </html>
